Add the original backup/restore from luddatumbazo
Originally some code for doing database backups and restores was added
to luddatumbazo; however the purpose of Sekurkopio is to be a more
generic solution.

This commit back ports the original code for this from luddatumbazo, with
some tweaks to make it more open ended (i.e. the database names are
parameters).